extract post creation transaction logic to service layer
move BEGIN/COMMIT/ROLLBACK and media processing out of the controller
into posts/service.ts — controller now calls service.createPhotoPost()
