+++ src/modules/posts/controller.ts
@@ -8,44 +8,14 @@ import { createModuleLogger } from '../../shared/utils/logger.js';
 import { NotFoundError, ValidationError } from '../../shared/errors/AppError.js';
 
 const log = createModuleLogger('posts');
-import path from 'path';
-import { fromBuffer as fileTypeFromBuffer } from 'file-type';
-import { Upload } from '@aws-sdk/lib-storage';
-import { DeleteObjectCommand } from '@aws-sdk/client-s3';
-import { s3, S3_BUCKET, CDN_BASE } from '../../config/s3.js';
-import {
-  type ProcessedImage,
-  type ProcessedVideo,
-  processImage,
-  processVideo,
-  ALLOWED_VIDEO_MIME,
-} from '../../shared/utils/mediaProcessor.js';
 import * as repo from './repository.js';
-import type { MediaRow } from './types.js';
+import * as service from './service.js';
 
 // ── Helpers ────────────────────────────────────────────────────────────────
 
 /** Express 5 params can be string | string[] */
 function param(val: string | string[]): string {
   return Array.isArray(val) ? val[0] : val;
-}
-
-async function s3Upload(
-  buffer: Buffer,
-  key: string,
-  contentType: string,
-): Promise<string> {
-  const up = new Upload({
-    client: s3,
-    params: {
-      Bucket: S3_BUCKET!,
-      Key: key,
-      Body: buffer,
-      ContentType: contentType,
-    },
-  });
-  await up.done();
-  return `${CDN_BASE}/${key}`;
 }
 
 // ── Multer ─────────────────────────────────────────────────────────────────
@@ -67,8 +37,7 @@ function runMulter(req: Request, res: Response): Promise<void> {
 }
 
 // ── Zod schema for createPost ──────────────────────────────────────────────
-// We inline-import the JS schema to avoid duplicating the definition.
-// eslint-disable-next-line @typescript-eslint/no-var-requires
+
 import { z } from 'zod';
 
 const createPostSchema = z.discriminatedUnion('type', [
@@ -151,90 +120,10 @@ export class PostsController {
         throw new ValidationError('Maximum 10 files allowed');
       }
 
-      const client = await (await import('../../config/database.js')).pool.connect();
       try {
-        await client.query('BEGIN');
-
-        const post = await repo.insertPhotoPost(sub, caption);
-        const postId = post.id;
-
-        const mediaRows: MediaRow[] = [];
-        for (let i = 0; i < files.length; i++) {
-          const fileBuffer = files[i].buffer;
-          const detected = await fileTypeFromBuffer(
-            fileBuffer.slice(0, 4100),
-          );
-
-          if (detected && ALLOWED_VIDEO_MIME.has(detected.mime)) {
-            // ── video file ─────────────────────────────────────────────────
-            let vidProcessed: ProcessedVideo;
-            try {
-              vidProcessed = await processVideo(fileBuffer);
-            } catch (vidErr: any) {
-              await client.query('ROLLBACK');
-              log.error({ err: vidErr }, 'video processing failed');
-              throw new ValidationError('Failed to process video');
-            }
-
-            const { thumbBuffer, width, height, duration } = vidProcessed;
-            const videoKey = `posts/${postId}/${i}_video${path.extname(files[i].originalname) || '.mp4'}`;
-            const thumbKey = `posts/${postId}/${i}_thumb.webp`;
-
-            const [videoUrl, thumbUrl] = await Promise.all([
-              s3Upload(fileBuffer, videoKey, detected.mime),
-              s3Upload(thumbBuffer, thumbKey, 'image/webp'),
-            ]);
-
-            const mediaRow = await repo.insertMediaRow(postId, {
-              s3Key: videoKey,
-              url: videoUrl,
-              width,
-              height,
-              position: i,
-              blurDataUrl: '',
-              mediaType: 'video',
-              thumbnailUrl: thumbUrl,
-              duration,
-            });
-            mediaRows.push(mediaRow);
-          } else {
-            // ── image file ─────────────────────────────────────────────────
-            let processed: ProcessedImage;
-            try {
-              processed = await processImage(fileBuffer);
-            } catch (imgErr: any) {
-              await client.query('ROLLBACK');
-              throw new ValidationError(imgErr.message);
-            }
-
-            const { fullBuffer, thumbBuffer, blurDataUrl, width, height } =
-              processed;
-            const fullKey = `posts/${postId}/${i}_full.webp`;
-            const thumbKey = `posts/${postId}/${i}_thumb.webp`;
-
-            const [fullUrl, thumbUrl] = await Promise.all([
-              s3Upload(fullBuffer, fullKey, 'image/webp'),
-              s3Upload(thumbBuffer, thumbKey, 'image/webp'),
-            ]);
-
-            const mediaRow = await repo.insertMediaRow(postId, {
-              s3Key: fullKey,
-              url: fullUrl,
-              width: width ?? 0,
-              height: height ?? 0,
-              position: i,
-              blurDataUrl,
-              mediaType: 'image',
-              thumbnailUrl: thumbUrl,
-            });
-            mediaRows.push(mediaRow);
-          }
-        }
-
-        await client.query('COMMIT');
-        return res.status(201).json({ ...post, media: mediaRows });
+        const result = await service.createPhotoPost(sub, caption, files);
+        return res.status(201).json(result);
       } catch (err: any) {
-        await client.query('ROLLBACK');
         if (err instanceof ValidationError) {
           next(err);
         } else {
@@ -242,8 +131,6 @@ export class PostsController {
           next(err);
         }
         return;
-      } finally {
-        client.release();
       }
     }
 
@@ -360,31 +247,7 @@ export class PostsController {
     const id = param(req.params.id);
 
     try {
-      // Grab S3 keys before delete
-      const mediaKeys = await repo.getMediaS3Keys(id);
-
-      const rowCount = await repo.deletePost(id, sub);
-      if (rowCount === 0) {
-        throw new NotFoundError('Post not found');
-      }
-
-      // Best-effort S3 cleanup
-      if (mediaKeys.length > 0) {
-        await Promise.allSettled(
-          mediaKeys.flatMap(({ s3_key }) => [
-            s3.send(
-              new DeleteObjectCommand({ Bucket: S3_BUCKET!, Key: s3_key }),
-            ),
-            s3.send(
-              new DeleteObjectCommand({
-                Bucket: S3_BUCKET!,
-                Key: s3_key.replace('_full.webp', '_thumb.webp'),
-              }),
-            ),
-          ]),
-        );
-      }
-
+      await service.deletePostWithMedia(id, sub);
       res.status(204).end();
     } catch (err: any) {
       next(err);
